backend/requirements.txt
# Core web framework
fastapi==0.110.1
uvicorn[standard]==0.25.0
starlette==0.37.2
python-multipart==0.0.20

# Database (PostgreSQL)
asyncpg==0.31.0

# Authentication & Security
PyJWT==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
cryptography>=46.0.0

# AI/LLM
openai>=1.0.0

# HTTP client
requests>=2.32.0
httpx>=0.28.0
httpcore>=1.0.0

# Data validation
pydantic==2.12.5
pydantic_core==2.41.5
email-validator>=2.3.0

# Environment & config
python-dotenv>=1.2.0

# Date/time utilities
python-dateutil>=2.9.0
pytz>=2025.2

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
